requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
SQLAlchemy==2.0.19
Flask-Login==0.6.2
Pillow==10.0.0
Flask-WTF==1.1.1
email-validator==2.0.0
Flask-Mail==0.9.1
PyJWT==2.8.0
Flask-SocketIO==5.3.6
python-socketio==5.8.0
Flask-Caching==2.1.0
redis==5.0.1
Flask-RESTful==0.3.10
flask-restx==1.3.0
